adicionando psin funcionando no hello universe

+++ arp/ip/psin/psin.h
@@ -67,7 +67,7 @@ public:
   }
 
   //Stored sin value.
-  uint32_t psin_memory;
+  float psin_memory;
 
   /**
    * Default constructor.
